python_service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
plotly==5.24.1
bokeh==3.6.3
pandas==2.2.3
numpy==2.2.1
python-dotenv==1.2.2
